requirements.txt
flask
opencv-python
torch
matplotlib
numpy
scikit-image
pillow
